pyproject.toml
[project]
name = "python-playground"
version = "0.1.0"
description = ""
authors = [
    {name = "unhurried",email = "unhurried@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dateutil (>=2.8,<3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
freezegun = "^1.5.1"

